#1671: Mark required fields in Swagger config
